requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
albucore==0.0.21
albumentations==1.4.23
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
comfy==0.0.1
-e git+https://github.com/WitchoutName/exceles-gradio-app@e37bb957ffbd789d7b916cdcaaf9eadb7072c3d1#egg=comfyui_controlnet_aux
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
embreex==2.17.7.post5
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.12.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==5.9.1
gradio_client==1.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iopath==0.1.10
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
manifold3d==3.0.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
proglog==0.1.10
protobuf==4.25.5
pycollada==0.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
Rtree==1.3.0
ruff==0.8.4
safehttpx==0.1.6
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.41.3
stringzilla==3.11.2
svg.path==6.3
svglib==1.5.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
trimesh==4.5.3
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
vhacdx==0.0.8.post1
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
zipp==3.21.0
